[MS-2] Refactored id's for entities. Fixed bug with changing users.

+++ src/main/java/com/doit/wheels/dao/entities/Guideline.java
@@ -40,8 +40,8 @@ public class Guideline extends Description {
 
     @Override
     public int hashCode() {
-        int result = this.getId();
+        long result = this.getId();
         result = 31 * result + getDescription().hashCode();
-        return result;
+        return (int) result;
     }
 }
